Unify utils and separate client helpers
