--- src/utils/exceptions.py
@@ -1,6 +1,0 @@
-class ClientError(Exception):
-    pass
-
-
-class ServerError(Exception):
-    pass
