--- src/utils/loggers.py
@@ -1,5 +1,0 @@
-import logging
-
-
-logging.basicConfig(level=logging.DEBUG, format="%(levelname)s LOG: %(message)s")
-logger = logging.getLogger(__name__)
